move TestStartWatcherFromCompactedRevision to common test package
