+++ tests/e2e/watch_test.go
@@ -25,13 +25,10 @@ import (
 	"testing"
 	"time"
 
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"golang.org/x/sync/errgroup"
 
-	"go.etcd.io/etcd/api/v3/etcdserverpb"
 	"go.etcd.io/etcd/api/v3/mvccpb"
-	v3rpc "go.etcd.io/etcd/api/v3/v3rpc/rpctypes"
 	clientv3 "go.etcd.io/etcd/client/v3"
 	"go.etcd.io/etcd/tests/v3/framework/e2e"
 )
@@ -249,163 +246,6 @@ func continuouslyExecuteGetAll(ctx context.Context, t *testing.T, g *errgroup.Gr
 		}
 		return nil
 	})
-}
-
-func TestStartWatcherFromCompactedRevision(t *testing.T) {
-	t.Run("compaction on tombstone revision", func(t *testing.T) {
-		testStartWatcherFromCompactedRevision(t, true)
-	})
-	t.Run("compaction on normal revision", func(t *testing.T) {
-		testStartWatcherFromCompactedRevision(t, false)
-	})
-}
-
-func testStartWatcherFromCompactedRevision(t *testing.T, performCompactOnTombstone bool) {
-	e2e.BeforeTest(t)
-	cfg := e2e.DefaultConfig()
-	cfg.Client = e2e.ClientConfig{ConnectionType: e2e.ClientTLS}
-	clus, err := e2e.NewEtcdProcessCluster(t.Context(), t, e2e.WithConfig(cfg), e2e.WithClusterSize(1))
-	require.NoError(t, err)
-	defer clus.Close()
-
-	c := newClient(t, clus.EndpointsGRPC(), cfg.Client)
-	defer c.Close()
-
-	ctx := t.Context()
-	key := "foo"
-	totalRev := 100
-
-	type valueEvent struct {
-		value string
-		typ   mvccpb.Event_EventType
-	}
-
-	var (
-		// requestedValues records all requested change
-		requestedValues = make([]valueEvent, 0)
-		// revisionChan sends each compacted revision via this channel
-		compactionRevChan = make(chan int64)
-		// compactionStep means that client performs a compaction on every 7 operations
-		compactionStep = 7
-	)
-
-	// This goroutine will submit changes on $key $totalRev times. It will
-	// perform compaction after every $compactedAfterChanges changes.
-	// Except for first time, the watcher always receives the compacted
-	// revision as start.
-	go func() {
-		defer close(compactionRevChan)
-
-		lastRevision := int64(1)
-
-		compactionRevChan <- lastRevision
-		for vi := 1; vi <= totalRev; vi++ {
-			var respHeader *etcdserverpb.ResponseHeader
-
-			if vi%compactionStep == 0 && performCompactOnTombstone {
-				t.Logf("DELETE key=%s", key)
-
-				resp, derr := c.KV.Delete(ctx, key)
-				assert.NoError(t, derr)
-				respHeader = resp.Header
-
-				requestedValues = append(requestedValues, valueEvent{value: "", typ: mvccpb.DELETE})
-			} else {
-				value := fmt.Sprintf("%d", vi)
-
-				t.Logf("PUT key=%s, val=%s", key, value)
-				resp, perr := c.KV.Put(ctx, key, value)
-				assert.NoError(t, perr)
-				respHeader = resp.Header
-
-				requestedValues = append(requestedValues, valueEvent{value: value, typ: mvccpb.PUT})
-			}
-
-			lastRevision = respHeader.Revision
-
-			if vi%compactionStep == 0 {
-				compactionRevChan <- lastRevision
-
-				t.Logf("COMPACT rev=%d", lastRevision)
-				_, err = c.KV.Compact(ctx, lastRevision, clientv3.WithCompactPhysical())
-				assert.NoError(t, err)
-			}
-		}
-	}()
-
-	receivedEvents := make([]*clientv3.Event, 0)
-
-	fromCompactedRev := false
-	for fromRev := range compactionRevChan {
-		watchChan := c.Watch(ctx, key, clientv3.WithRev(fromRev))
-
-		prevEventCount := len(receivedEvents)
-
-		// firstReceived represents this is first watch response.
-		// Just in case that ETCD sends event one by one.
-		firstReceived := true
-
-		t.Logf("Start to watch key %s starting from revision %d", key, fromRev)
-	watchLoop:
-		for {
-			currentEventCount := len(receivedEvents)
-			if currentEventCount-prevEventCount == compactionStep || currentEventCount == totalRev {
-				break
-			}
-
-			select {
-			case watchResp := <-watchChan:
-				t.Logf("Receive the number of events: %d", len(watchResp.Events))
-				for i := range watchResp.Events {
-					ev := watchResp.Events[i]
-
-					// If the $fromRev is the compacted revision,
-					// the first event should be the same as the last event receives in last watch response.
-					if firstReceived && fromCompactedRev {
-						firstReceived = false
-
-						last := receivedEvents[prevEventCount-1]
-
-						assert.Equalf(t, last.Type, ev.Type,
-							"last received event type %s, but got event type %s", last.Type, ev.Type)
-						assert.Equalf(t, string(last.Kv.Key), string(ev.Kv.Key),
-							"last received event key %s, but got event key %s", string(last.Kv.Key), string(ev.Kv.Key))
-						assert.Equalf(t, string(last.Kv.Value), string(ev.Kv.Value),
-							"last received event value %s, but got event value %s", string(last.Kv.Value), string(ev.Kv.Value))
-						continue
-					}
-					receivedEvents = append(receivedEvents, ev)
-				}
-
-				if len(watchResp.Events) == 0 {
-					require.Equal(t, v3rpc.ErrCompacted, watchResp.Err())
-					break watchLoop
-				}
-
-			case <-time.After(10 * time.Second):
-				t.Fatal("timed out getting watch response")
-			}
-		}
-
-		fromCompactedRev = true
-	}
-
-	t.Logf("Received total number of events: %d", len(receivedEvents))
-	require.Len(t, requestedValues, totalRev)
-	require.Lenf(t, receivedEvents, totalRev, "should receive %d events", totalRev)
-	for idx, expected := range requestedValues {
-		ev := receivedEvents[idx]
-
-		require.Equalf(t, expected.typ, ev.Type, "#%d expected event %s", idx, expected.typ)
-
-		updatedKey := string(ev.Kv.Key)
-
-		require.Equal(t, key, updatedKey)
-		if expected.typ == mvccpb.PUT {
-			updatedValue := string(ev.Kv.Value)
-			require.Equal(t, expected.value, updatedValue)
-		}
-	}
 }
 
 // TestResumeCompactionOnTombstone verifies whether a deletion event is preserved
